refactor: more stuff
